changed structure of folders and added base html
