+++ templates/index.html
@@ -1,34 +1,6 @@
-<!DOCTYPE html>
-<html lang="en">
+{% extends "base.html" %}
 
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>FitPal - Your Personal Fitness Companion</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
-</head>
-
-<body>
-    <!-- Navigation Bar -->
-    <nav class="navbar">
-        <div class="logo">FitPal</div>
-        <div class="nav-links">
-            <a href="{{ url_for('index') }}#home">Home</a>
-            <a href="{{ url_for('index') }}#features">Features</a>
-            <a href="{{ url_for('workout_tools') }}">Tools</a>
-            <a href="{{ url_for('posts') }}">Posts</a>
-            <a href="{{ url_for('index') }}#about">About</a>
-            <a href="{{ url_for('index') }}#contact">Contact</a>
-            <a href="{{ url_for('login') }}">
-                <button class="login-btn">Login</button>
-            </a>
-            <a href="{{ url_for('signup') }}">
-                <button class="signup-btn">Sign Up</button>
-            </a>   
-        </div>
-    </nav>
-
+{% block content %}
     <!-- Hero Section -->
     <section id="home" class="hero">
         <div class="hero-content">
@@ -92,36 +64,4 @@
             <button type="submit">Send Message</button>
         </form>
     </section>
-
-    <!-- Footer -->
-    <footer>
-        <div class="footer-content">
-            <div class="footer-section">
-                <h3>FitPal</h3>
-                <p>Your personal fitness companion</p>
-            </div>
-            <div class="footer-section">
-                <h3>Quick Links</h3>
-                <a href="{{ url_for('index') }}#home">Home</a>
-                <a href="{{ url_for('index') }}#features">Features</a>
-                <a href="{{ url_for('index') }}#about">About</a>
-                <a href="{{ url_for('index') }}#contact">Contact</a>
-            </div>
-            <div class="footer-section">
-                <h3>Connect With Us</h3>
-                <div class="social-links">
-                    <a href="#"><i class="fab fa-facebook"></i></a>
-                    <a href="#"><i class="fab fa-twitter"></i></a>
-                    <a href="#"><i class="fab fa-instagram"></i></a>
-                </div>
-            </div>
-        </div>
-        <div class="footer-bottom">
-            <p>&copy; 2024 FitPal. All rights reserved.</p>
-        </div>
-    </footer>
-
-    <script src="{{ url_for('static', filename='script.js') }}"></script>
-</body>
-
-</html>
+{% endblock %}
