+++ templates/login.html
@@ -1,29 +1,6 @@
-<!DOCTYPE html>
-<html lang="en">
+{% extends "base.html" %}
 
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
-
-    <script src="{{ url_for('static', filename='script.js') }}" defer></script>
-</head>
-
-<body>
-    <nav class="navbar">
-        <div class="logo">FitPal</div>
-        <div class="nav-links">
-            <a href="{{ url_for('index') }}">Home</a>
-            <a href="{{ url_for('index') }}#features">Features</a>
-            <a href="{{ url_for('workout_tools') }}">Tools</a>
-            <a href="{{ url_for('posts') }}" class="active">Posts</a>
-            <a href="{{ url_for('index') }}#about">About</a>
-            <a href="{{ url_for('index') }}#contact">Contact</a>
-        </div>
-    </nav>
-
+{% block content %}
     <div class="login-container">
         <form action="/submit_form" method="POST" class="login-form">
             <h2>Login</h2>
@@ -58,5 +35,4 @@
             </div>
         </form>
     </div>
-</body>
-</html>
+{% endblock %}
